backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
alembic==1.13.2
email-validator>=2.0.0
requests>=2.32.0
pdfminer.six>=20231228
pdfplumber>=0.11.0
pypdf>=4.2.0
pyodbc>=4.0.39
openai>=1.51.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
